Putative Meetup inclusion
This might not be the best way to do it, but it does seem to get
the job done... It could be a start, but someone else probably has
a better eye than me!

It looks like a complement to https://github.com/toronto-perl-mongers/tpm-website/issues/5
as each meeting could have its own RSVP widget (or maybe the layout for
meetings "in the future" would replace the generic widget with an RSVP
one).